Desktop/Projects/Helping-Hand-master/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
Django==3.1.14
# django-allauth==0.46.0
django-environ==0.8.1
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.3.1
urllib3==1.26.7
